pyproject.toml
[project]
name = "noto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fonttools>=4.60.1",
    "ipython>=9.7.0",
    "pyinstrument>=5.1.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
